third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-events.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup events</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popup-utils.js"></script>

<div popup>Popup</div>

<script>
window.onload = () => {
  for(const method of ["listener","attribute"]) {
    promise_test(async t => {
      const popup = document.querySelector('[popup]');
      assert_false(popup.matches(':open'));
      let showCount = 0;
      let hideCount = 0;
      switch (method) {
        case "listener":
          const controller = new AbortController();
          const signal = controller.signal;
          t.add_cleanup(() => controller.abort());
          document.addEventListener('popupshow',() => ++showCount, {signal});
          document.addEventListener('popuphide',() => ++hideCount, {signal});
          break;
        case "attribute":
          assert_false(popup.hasAttribute('onpopupshow'));
          assert_false(popup.hasAttribute('onpopuphide'));
          t.add_cleanup(() => popup.removeAttribute('onpopupshow'));
          t.add_cleanup(() => popup.removeAttribute('onpopuphide'));
          popup.onpopupshow = () => ++showCount;
          popup.onpopuphide = () => ++hideCount;
          break;
        default: assert_unreached();
      }
      assert_equals(0,showCount);
      assert_equals(0,hideCount);
      popup.showPopUp();
      assert_true(popup.matches(':open'));
      assert_equals(1,showCount);
      assert_equals(0,hideCount);
      await waitForRender();
      assert_true(popup.matches(':open'));
      popup.hidePopUp();
      assert_false(popup.matches(':open'));
      assert_equals(1,showCount);
      assert_equals(1,hideCount);
      await waitForRender();
      // No additional events after animation frame
      assert_false(popup.matches(':open'));
      assert_equals(1,showCount);
      assert_equals(1,hideCount);
    }, `Popupshow and popuphide events (${method}) get properly dispatched for popups`);
  }

  promise_test(async t => {
    const popUp = document.querySelector('[popup]');
    const controller = new AbortController();
    const signal = controller.signal;
    t.add_cleanup(() => controller.abort());
    let cancel = true;
    popUp.addEventListener('popupshow',(e) => {
      if (cancel)
        e.preventDefault();
    }, {signal});
    assert_false(popUp.matches(':open'));
    popUp.showPopUp();
    assert_false(popUp.matches(':open'),'The "popupshow" event should be cancelable');
    cancel = false;
    popUp.showPopUp();
    assert_true(popUp.matches(':open'));
    popUp.hidePopUp();
    assert_false(popUp.matches(':open'));
  }, 'Popupshow event is cancelable');
};
</script>
